+++ inst/NEWS.md
@@ -1,8 +1,8 @@
 # Version 1.8.2 (2016-xx-xx)
 * `*Matrix` and `*Array` now additionally allow to check for integerish storage
   type via argument "mode".
-* Functions `*Count`, `*Int`, `*Number`, `*Integer` and `*Numeric` do not accept
-  logical values any more. `*Integerish` still does.
+* Functions `*Count`, `*Int`, `*Number`, `*Integer`, `*Integerish` and
+  `*Numeric` do not accept logical values any more.
 * The set functions are now more restrict regarding the class, e.g. distinguish
   between factors and characters.